requirements.txt
dill
tqdm
numpy
pytest-benchmark
torchtext
spacy
